feat: add exception

+++ codec/parser.py
@@ -40,8 +40,8 @@ class Parser(object):
     def get_lora_train_num(self) -> str:
         return self.config["lora"]["train_num"]
 
-    def get_loss_decided_num(self) -> int:
-        return int(self.config["lora"]["loss_decide"])
+    def get_loss_decided_num(self) -> float:
+        return float(self.config["lora"]["loss_decide"])
 
     def get_sd_service_addr(self) -> str:
         return str(self.config["child_service"]["sd"])
